Changed structure of dotfile directory to enable use of DFM
The dotfiles are now located in the home directory, in the exact structure that they would be in home. Rather than using stow to install the symlinks, I have inluded dfm in the repo. DFM can be found at https://github.com/justone/dotfiles. This utility allows for the quick installation of all dotfiles. The .dfminstall file provides configuration instructions.